add actions to list view
