+++ src/components/BookList.tsx
@@ -26,6 +26,8 @@ import { fetchCollection } from "dataflow/opds1/fetch";
 import { useSWRInfinite } from "swr";
 import ApplicationError from "errors";
 import useUser from "components/context/UserContext";
+import FulfillmentButton from "components/FulfillmentButton";
+import { dedupeLinks, getFulfillmentDetails } from "utils/fulfill";
 
 const ListLoadingIndicator = () => (
   <div
@@ -258,6 +260,14 @@ const BookListCTA: React.FC<{ book: AnyBook }> = ({ book }) => {
         ? `You have this book on loan until ${availableUntil}.`
         : "You have this book on loan.";
 
+    // we will show a fulfillment button if there is only one option
+    const dedupedLinks = dedupeLinks(book.fulfillmentLinks);
+    const fulfillments = dedupedLinks
+      .map(getFulfillmentDetails)
+      .filter(details => details.type !== "unsupported");
+    const singleFulfillment =
+      fulfillments.length === 1 ? fulfillments[0] : undefined;
+
     return (
       <>
         <Text
@@ -266,6 +276,13 @@ const BookListCTA: React.FC<{ book: AnyBook }> = ({ book }) => {
         >
           {subtitle}
         </Text>
+        {singleFulfillment && (
+          <FulfillmentButton
+            details={singleFulfillment}
+            book={book}
+            isPrimaryAction
+          />
+        )}
         <NavButton variant="ghost" bookUrl={book.url} iconRight={ArrowForward}>
           View Book Details
         </NavButton>
