--- src/hooks/useIsMounted.ts
@@ -1,14 +1,0 @@
-import * as React from "react";
-
-export default function useIsMountedRef() {
-  const isMounted = React.useRef(false);
-
-  React.useEffect(() => {
-    isMounted.current = true;
-    return () => {
-      isMounted.current = false;
-    };
-  }, []);
-
-  return isMounted;
-}
